feat: toggle editor on click
