+++ src/components/index.js
@@ -5,3 +5,4 @@ export { SideBar } from "./SideBar/SideBar";
 export { FilterModal } from "../components/Modals/FilterModal/Filtermodal.jsx";
 export { ColorPallete } from "../components/Modals/ColorPalleteModal/ColorPallete.jsx";
 export { LabelModal } from "../components/Modals/LabelModal/LabelModal.jsx";
+export { Editor } from "../components/Editor/Editor";
